fully made restful api (only one-to-many)

+++ src/main/java/eat_it_server/repository/AdditivesRepository.java
@@ -3,5 +3,5 @@ package eat_it_server.repository;
 import eat_it_server.model.Additives;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdditivesRepository extends JpaRepository<Additives, Integer> {
+public interface AdditivesRepository extends JpaRepository<Additives, Long> {
 }
